#1 refactor: 색상 선택 및 옵션 선택 관련 컴포넌트 파일 분리
