+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/component/ColorSelectItems.kt
@@ -1,0 +1,100 @@
+package com.softeer.mycarchiving.ui.component
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.softeer.mycarchiving.R
+import com.softeer.mycarchiving.ui.theme.HyundaiActiveBlue
+import com.softeer.mycarchiving.ui.theme.roundCorner
+
+@Composable
+fun CarImageSelectItem(
+    modifier: Modifier,
+    onItemClick: () -> Unit
+) {
+    var itemSelect by remember { mutableStateOf(false) }
+    Surface(
+        modifier = modifier
+            .size(80.dp)
+            .clickable {
+                itemSelect = !itemSelect
+                onItemClick()
+            },
+        shape = roundCorner,
+        border = if (itemSelect) BorderStroke(width = 2.dp, color = HyundaiActiveBlue) else null
+    ) {
+        Image(
+            modifier = modifier
+                .fillMaxSize(),
+            painter = painterResource(id = R.drawable.ic_launcher_background),
+            contentDescription = null,
+            contentScale = ContentScale.Crop
+        )
+    }
+}
+
+@Composable
+fun CarColorSelectItem(
+    modifier: Modifier,
+    onItemClick: () -> Unit
+) {
+    var itemSelect by remember { mutableStateOf(false) }
+    Box(
+        modifier = modifier
+            .size(59.dp)
+            .clickable {
+                itemSelect = !itemSelect
+                onItemClick()
+            }
+    ){
+        Surface(
+            modifier = modifier
+                .size(48.dp)
+                .align(Alignment.Center),
+            shape = roundCorner,
+            border = if (itemSelect) BorderStroke(width = 3.dp, color = HyundaiActiveBlue) else null
+        ) {
+            Image(
+                modifier = modifier
+                    .fillMaxSize(),
+                painter = painterResource(id = R.drawable.ic_launcher_background),
+                contentDescription = null,
+                contentScale = ContentScale.Crop
+            )
+        }
+        if (itemSelect) {
+            CheckCircle(
+                modifier = modifier
+                    .align(Alignment.TopEnd)
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun PreviewCarImageSelectItem() {
+    CarImageSelectItem(modifier = Modifier, onItemClick = {})
+}
+
+
+@Preview
+@Composable
+fun PreviewCarColorSelectItem() {
+    CarColorSelectItem(modifier = Modifier, onItemClick = {})
+}
