+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/component/OptionSelectItem.kt
@@ -6,6 +6,8 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -13,6 +15,9 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.AssistChip
+import androidx.compose.material3.AssistChipDefaults
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -34,6 +39,7 @@ import com.softeer.mycarchiving.ui.theme.HyundaiLightSand
 import com.softeer.mycarchiving.ui.theme.PrimaryBlue
 import com.softeer.mycarchiving.ui.theme.PrimaryBlue10
 import com.softeer.mycarchiving.ui.theme.medium14
+import com.softeer.mycarchiving.ui.theme.regular14
 import com.softeer.mycarchiving.ui.theme.roundCorner
 
 @Composable
@@ -99,78 +105,48 @@ fun PreviewOptionSelectItem() {
     OptionSelectItem(modifier = Modifier, optionName = "컴포트 2", optionPrice = "1,090,000", onItemClick = {}, onAddClick = {})
 }
 
+@OptIn(ExperimentalLayoutApi::class)
 @Composable
-fun CarImageSelectItem(
-    modifier: Modifier,
-    onItemClick: () -> Unit
+fun OptionSelectedInfo(
+    optionTags: List<String>
 ) {
-    var itemSelect by remember { mutableStateOf(false) }
-    Surface(
-        modifier = modifier
-            .size(80.dp)
-            .clickable {
-                itemSelect = !itemSelect
-                onItemClick()
-            },
-        shape = roundCorner,
-        border = if (itemSelect) BorderStroke(width = 2.dp, color = HyundaiActiveBlue) else null
+    OptionHeadText(optionName = "컴포트 II")
+    OptionHeadComment()
+    FlowRow(
+        verticalAlignment = Alignment.Top,
+        horizontalArrangement = Arrangement.spacedBy(4.dp)
     ) {
-        Image(
-            modifier = modifier
-                .fillMaxSize(),
-            painter = painterResource(id = R.drawable.ic_launcher_background),
-            contentDescription = null,
-            contentScale = ContentScale.Crop
+        optionTags.forEach { tagString ->
+            OptionTagChip(tagString = tagString)
+        }
+    }
+}
+
+@Preview(heightDp = 140)
+@Composable
+fun PreviewOptionSelectedInfo() {
+    Column(
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        OptionSelectedInfo(
+            optionTags = listOf(
+                "어린이🧒",
+                "안전사고 예방🚨",
+                "대형견도 문제 없어요🐶",
+                "가족들도 좋은 옵션👨‍👩‍👧‍👦"
+            )
         )
     }
 }
 
-@Preview
-@Composable
-fun PreviewCarImageSelectItem() {
-    CarImageSelectItem(modifier = Modifier, onItemClick = {})
-}
 
+@Preview(heightDp = 60)
 @Composable
-fun CarColorSelectItem(
-    modifier: Modifier,
-    onItemClick: () -> Unit
-) {
-    var itemSelect by remember { mutableStateOf(false) }
-    Box(
-        modifier = modifier
-            .size(59.dp)
-            .clickable {
-                itemSelect = !itemSelect
-                onItemClick()
-            }
-    ){
-        Surface(
-            modifier = modifier
-                .size(48.dp)
-                .align(Alignment.Center),
-            shape = roundCorner,
-            border = if (itemSelect) BorderStroke(width = 3.dp, color = HyundaiActiveBlue) else null
-        ) {
-            Image(
-                modifier = modifier
-                    .fillMaxSize(),
-                painter = painterResource(id = R.drawable.ic_launcher_background),
-                contentDescription = null,
-                contentScale = ContentScale.Crop
-            )
-        }
-        if (itemSelect) {
-            CheckCircle(
-                modifier = modifier
-                    .align(Alignment.TopEnd)
-            )
-        }
+fun PreviewOptionColorNameSentence() {
+    Column(
+        verticalArrangement = Arrangement.SpaceBetween
+    ) {
+        OptionHeadText(optionName = "퀄팅 천연(블랙)")
+        OptionHeadComment()
     }
-}
-
-@Preview
-@Composable
-fun PreviewCarColorSelectItem() {
-    CarColorSelectItem(modifier = Modifier, onItemClick = {})
 }
